Move molecular calculations in test_ham_generalized_chemical
